Getting Windows to play nice with this setup
