+++ old/BucketSerializer.ts
