+++ package.json
@@ -27,10 +27,10 @@
     "typescript": "^3.6.3"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.18",
+    "@types/jest": "^24.9.1",
     "@types/node": "^13.9.0",
     "jest": "^25.1.0",
-    "ts-jest": "^24.1.0",
+    "ts-jest": "^25.2.1",
     "tslint": "^6.0.0"
   }
 }
