+++ src/errors/MissingBucketError.ts
@@ -1,5 +1,3 @@
-import Bucket from "../bucket";
-
 // This works where other, more comprehensible custom errors don't.
 
 export class MissingBucketError {
